Move runVisualDiffs and related modules into firefox subfolder
This module is specific to the firefox target. By moving everything
firefox related into a separate folder, we're making the upcoming
transition to a plugin framework easier.
