+++ src/server/firefox/closeDriver.js
