+++ src/server/firefox/initializeWebdriver.js
@@ -1,8 +1,8 @@
-require('./firefox/addGeckoDriverToPath');
+require('./addGeckoDriverToPath');
 
 const seleniumWebdriver = require('selenium-webdriver');
 
-const { config } = require('./config');
+const { config } = require('../config');
 
 module.exports = function initializeWebdriver() {
   return new Promise((resolve, reject) => {
